# multiplayer-shooter-api/build/swagger.json
{
    "components": {
...
        "schemas": {
...
            "IResponseServerError": {
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "error": {
                        "properties": {},
                        "additionalProperties": {
                            "type": "string"
                        },
                        "type": "object"
                    }
                },
                "required": [
                    "message",
                    "error"
                ],
                "type": "object",
                "additionalProperties": false
            },
...
            "IResponseErrors": {
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "items": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "field": {
                                    "type": "string"
                                }
                            },
                            "required": [
                                "message",
                                "field"
                            ],
                            "type": "object"
                        },
                        "type": "array"
                    }
                },
                "required": [
                    "message",
                    "errors"
                ],
                "type": "object",
                "additionalProperties": false
            },
...
            "IMapJSON": {
                "properties": {
                    "_id": {
                        "type": "string"
                    },
                    "mapTitle": {
                        "type": "string"
                    },
                    "mapTexturePath": {
                        "type": "string"
                    },
                    "mapCollisionTexturePath": {
                        "type": "string"
                    },
                    "mapCreatedAt": {
                        "type": "string"
                    },
                    "mapUpdatedAt": {
                        "type": "string"
                    }
                },
                "required": [
                    "_id",
                    "mapTitle",
                    "mapTexturePath",
                    "mapCollisionTexturePath",
                    "mapCreatedAt",
                    "mapUpdatedAt"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "IResponse_Array_IMapJSON__": {
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "items": {
                            "$ref": "#/components/schemas/IMapJSON"
                        },
                        "type": "array"
                    }
                },
                "required": [
                    "message"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "IMapCreate": {
                "properties": {
                    "mapTitle": {
                        "type": "string"
                    }
                },
                "type": "object",
                "additionalProperties": false
            },
            "IResponse_IMapJSON_": {
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/IMapJSON"
                    }
                },
                "required": [
                    "message"
                ],
                "type": "object",
                "additionalProperties": false
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "access_token",
                "in": "query"
            }
        }
    },
    "info": {
        "title": "shooter-api",
        "version": "1.0.0",
        "license": {
            "name": "ISC"
        },
        "contact": {}
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/maps": {
            "get": {
                "operationId": "GetMaps",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IResponse_Array_IMapJSON__"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IResponseServerError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": []
            },
            "post": {
                "operationId": "PostImage",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IResponse_IMapJSON_"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IResponseErrors"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IResponseServerError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": [
                            "admin"
                        ]
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IMapCreate"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}